refactor(client): P8 — autonomous-loop unification onto ToolLoopRunner (#140)
* refactor(client): P8(1/2) — autonomous loop onto ToolLoopRunner, duplicate deleted

_run_loop_iteration (308 lines) moves verbatim into ToolLoopRunner.run_
autonomous (host-based); _dispatch_loop_tool(+inner) become runner
methods; _LoopMessageProxy/_LoopAuthorProxy move to tool_loop.py with a
client re-export alias (tests import them from client). Bot keeps
delegates for all three (api.py calls _run_loop_iteration; the scheduled
path and tests call the dispatch pair).

Both pipelines now live in one module sharing one dispatch table. Every
sec-4.3 asymmetry pin green: CircuitOpenError re-raise, no guards/
classifier on the loop path, raw llm_client seam (no gateway wiring),
cap -> is_error/failure_class=cancelled with stale-partial surfacing,
recovered-error reflection semantics, export staging, proxy dispatch.

* refactor(client): P8(2/2) — LoopPolicy + shared-fragment extraction

- LoopPolicy dataclass with CHAT_POLICY/AUTONOMOUS_POLICY constants: the
  sec-4.3 dimensions as executable data-of-record. Load-bearing wirings:
  run() reads policy.skill_file_delivery for skill dispatch and
  policy.trajectory_source for the turn source; run_autonomous() reads
  policy.trajectory_source. The remaining fields document the divergences
  that live in the two entry points' control flow, with an explicit
  do-not-unify-without-a-pin warning.
- build_assistant_content(): the byte-identical assistant tool_use
  fragment both pipelines carried, extracted once (R5 seam work).
- Declared test migration (first of the campaign): _LoopIterClient in
  test_trajectory_completeness rebinds the real _run_loop_iteration onto
  a fake host — post-P8 that method delegates to the host-based runner,
  so the fake now constructs its own ToolLoopRunner(self). The test still
  drives the REAL loop body; assertions unchanged.
- ruff clean on tool_loop.py.
